safari-600.7-branch/LayoutTests/mathml/presentation/inferred-mrow-baseline-expected.txt

This test passes if rectangles are correctly aligned with respect to the baseline.

 
 
 
 
 
 

Mrow: PASS
Sqrt: PASS
Style: PASS
Error: PASS
Phantom: PASS
Math: PASS
Menclose: PASS
